add specificity to the operators
ugh the unary expressions, I don't want to move ~ but it looks bad

Some operators are used in more than one scenario, but only have one operation scope, so I still put it in. For example /, <.
Not adding scopes for the words or the multi-operation operators. For example instanceof, +

I put them in even if they're used in only 1 scope. So any short files like this:
>>

Won't colorize because it's out of scope
Maybe add 'invalid.keyword.operator.js'